[new-env] Let the env builder use the context
Summary: It will be nice to be able to directly modify the context from the env builder, for example, to raise TDZ errors. This diff changes the dependencies so that the env builder can use some modules in /typing; it also creates several modules representing `Flow_js` and `Context` that can be used from within the env_builder. (A cyclical dependency prevents the env_builder from directly using `Flow_js` unless we moved it into /typing). A lot of the type definitions and helper functions within the env builder are moved to the env_api, which is probably the correct thing anyways.

Reviewed By: jbrown215

Differential Revision: D29979193

fbshipit-source-id: f1c5a09c4409da900bfe8094d0794b2a85258699