[new-env] Create loc-keyed environment data structure
Summary:
Previously known as ssa_env.ml, this diff creates a data structure mapping locs to types for use as the core type storage structure in the new environment. (It's a record because new entries may need to be added in the future.)

Previously this data structure also contained the output of the env builder; that is also stored elsewhere in the context so this version separates those concerns.

Reviewed By: samwgoldman

Differential Revision: D29857272

fbshipit-source-id: e52808f6b6135384f9d91610b2eafeb7ac865c1d