[new-env] Include environment work in context
Summary: This diff creates a context entry for the `var_env`, which was defined in the previous diff in this stack. It also modifies how the result of the env_builder is stored in the context--previously it was optional and called `use_def`, but now it's necessary for any typechecking so it's no longer optional; and its also no longer most important as the use-def chain but instead encapsulates all info about variable resolution, so it's renamed `var_info`.

Reviewed By: samwgoldman

Differential Revision: D29857496

fbshipit-source-id: 513957fa47b800a30b21cff177c8a7e2840d4030