Store workload items as records
Summary: This diff doesn't change anything about the observable behavior or interface of the `WorkdloadStream` data structure, it just rewrites some tuples as records in preparation for adding a `name` field to the record to track names of workload items. For consistency, this diff adds storing of enqueue times for requeued items.

Reviewed By: mroch

Differential Revision: D30560330

fbshipit-source-id: a14c55524e8fcfc16176cbc125f527b7304556a9