track unique names for server workloads
Summary: This diff makes the `WorkloadStream` data structure take a name for each item it is asked to enqueue, so that in the next diff we can include this name for logging. A unique name is passed for each request, both for ephemeral and LSP requests. In ephemeral requests, the request ID used in our custom Monitor RPC is used. In LSP requests, the JSON-RPC ID of the request is used.

Reviewed By: mroch

Differential Revision: D30560331

fbshipit-source-id: 51bb60077d376f6d0e66b76f00bc397440743823