codemod runner uses Job_utils
Summary:
This helps prevent some timeouts when running the codemod on large repos.
This was due to the previous version not using the timeouts and caching facilities
that Job_utils offers.

One internal change in the codemod runner is that now the result of the check
pass is a list of "file * result" instead of a map from files to results. This should not
have an effect on the final outcome.

Changelog: [fix] improve codemod performance on large codebases

Reviewed By: evanyeung

Differential Revision: D30687266

fbshipit-source-id: 084a1649f7ea6a072a9fecea668f007b8efd915f