[new-env] Create handler module for the new environment and implement variable initialization
Summary:
This diff creates the main interface module for the new environment, which will handle all writes/reads/refinements/etc. Currently, it only contains the implementation for initializing the environment, by looking at all env builder values and populating the `var_env` with a tvar representing them (conceptually corresponding to the `statement_decl` pass.) It also implements some helper functions to be used later on.

One unimplemented caveat is that this means that variable assignments that are never directly read from are not initialized--for example,
```
var x = 10
x = 42
```
will only create an entry for `x = 42`. This will be handled later in the env builder.

Reviewed By: panagosg7

Differential Revision: D29858166

fbshipit-source-id: 53598212320e2491617d22090e195556ee9d4698