[new-env] Implement reading and refining in new environment
Summary:
This is the key rad work dsainati1 did previously in D28412985 updated to work with the new model of the env builder.

When we read a variable, we find a set of env builder writes, which can be locations, uninitialized, or another set of writes that have been refined. We look up locations in the var env, and if they're part of a refined set we join them into a tvar, refine that tvar with predicates, and flow the result into the overall type. this recursive process generates an overall, summarized view of a variable.

As a side note, this diff includes some verbose-debug printing, but since there isn't a trace available here it also makes traces optional for debug printing, which causes minor changes in lots of places.

Reviewed By: jbrown215

Differential Revision: D29858582

fbshipit-source-id: 2eb078000ed672c8d7cabdf91c5a9fb10b5529ee