annotation constraint graph
Summary:
A new constraint graph for tvars created during annotation inference. This graph
is implemented with the same union-find structure used for the main type graph.

## Annotation constraints

Unlike `Type.Constraint` that may gradually evolve towards the solution of an
unresolved tvar, annotation constraints are resolved immediately.

There are three kinds of constraints that correspond to different states
of knowledge about a type annotation:

1/ `Annot_unresolved` is a constraint for which we have no information yet.
This state is associated, for example, with a type variable representing a type
alias that we have just started visiting, and so do not have a type for its body
yet.

2/ `Annot_op (_, op, id)` expresses the fact that the current variable is the
result of applying the operation `op` on the type that another annotation variable
`id` will resolve to.

3/ `Annot_resolved` is the state of a fully-resolved annotation variable.

## Transitions

An annotation variable starts off in one of the first two states and is resolved
to an Annot_resolved constraint in one step. For the first two kinds of constraints
we keep track of other tvars that depend on the result of the current constraint.
Once the tvar is resolved, the dependent set of the variable is forced to evaluation.
(See `Annotation_inference.resolved_id`.)

While in one of the two initial states the respective annotation variable ids are
only present in the annotation graph. Once resolved, the resolved type is immediately
stored in the main type graph as `FullyResolved` constraint. This is important
when looking up type variable ids during annotation inference.

## A note on laziness

Lazy unresolved constraints (D27382486 (https://github.com/facebook/flow/commit/e5bcc1b3ffaf61a40218c12f753a2b025dd30b0f)) can get a bit tricky in this new mode,
because an unresolved annotation constraint is not representable in the (classic)
type graph. (This is a key point of the design.)

This issue comes up in porting `mk_lazy_tvar` to annotation_inference.ml. This
function works very similarly to the check_service.ml counter-part: Its return
expression is a lazy, yet fully-resolved constraint (as eventually all annotations
are fully resolved).

The "constraint" thunk itself (let's call it `c`) starts off with an unresolved *annot* tvar
and applies a function `f` which is passed as argument. This is typically an operation
that computes the final result.

The expectation is that by the time `c` is forced, the relevant annotation constraint `a`
 will have boiled away, leaving behind `FullyResolved` type constraints.

A problematic situation may arise if constraint `c` gets forced *before* all annotation
constraints have been resolved. This is the case, for example, when we try to force
`c` from within annotation_inference.ml. (See D29180014 for a specific example of this.)

When this happens we have to produce a value for `c` based on the form of `a` at
that point. If `a` is `Annot_resolved` then the choice is easy. In any other case, we
consider the annotation constraints stuck and produce `any` as the resulting type.

## Well-formedness check

While empirically the above expectation holds for well-formed programs, just like in
the existing system, this is not a guarantee.

In rare cases like
```
// file rec-export.js
import {p} from './rec-export';
export {p};
```
it is possible for a variable to never become resolved. This is the equivalent
of a type variable in the old style of constraints to never accumulate any lower
bounds.

There are a few places in the implementation where we can detect such cases. A
convenient one is in `mk_lazy_tvar`, since without a fall-back type we wouldn't be
able to produce a `FullyResolved` constraint.

If at the point where this constraint is forced, we are still in an unresolved annotation,
then the result becomes a fully-resolved `any`.

Note that today, this case is detected by looking at the lower bounds of an unresolved
tvar during the copier stage.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D28496902

fbshipit-source-id: 9ed8168fca5bb1b4ee7a1690599e4c3e11fef267