annotate-empty-object should add indexer annotations on literal indexed property writes
Summary:
When Flow sees a write like the following:

```
declare var u : "A" | "B" | "C";
let o = {};
o[u] = 3;
```

we do not process it internally with the indexed write machinery. To see this, look at the hover type of `o` in a TryFlow example.

We still need to add annotations on `o` here, however, in order to get rid of unsealed empty objects, so this modifies the codemod to handle this case.

Reviewed By: panagosg7

Differential Revision: D30731345

fbshipit-source-id: d509d57485b4679f7864c3e756c25863e94a7d78