fix unused suppression errors in lazy mode
Summary:
in lazy mode, type errors should only be raised for checked files -- files in `CheckedSet`. but there's a recheck opt that skips checking a dependent if its dependencies' signatures haven't changed, even if it's in the `CheckedSet`.

the file could still be merged if it hasn't been seen yet. merging the file adds its suppressions to the suppression set, but since we skip checking it, we never see the errors that are meant to be suppressed so we get spurious unused suppression warnings.

to fix, we ignore the suppressions from skipped dependents.

in turn, this causes suppressed signature verification errors -- which are found during parsing -- to start erroring, because we ignored their suppressions! since we want sig errors to only show up during check, I moved them into the `FileSigHeap` and add them to `merge_errors` during check.

Changelog: [fix] Fix a bug that could cause "unused suppression" warnings to appear incorrectly when using lazy mode

Reviewed By: panagosg7

Differential Revision: D30740729

fbshipit-source-id: a13550ff54aaed353b8245f435aea13248ec9dd3