[new_website] Migrate the rest of the docs
Summary:
This diff migrates the final 4 categories to the new website. It was
the same process as in the other diffs of this stack (update top of page
metadata and header links).

There isn't an automatic way to create an index page like at
https://flow.org/en/docs/tools/ in docusaurus. It looks like there's an issue
someone is working on to add this feature
https://github.com/facebook/docusaurus/issues/2643. However, I just skipped the
index. I don't think it's necessary for our docs and we don't use this pattern
in any of the other sections.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D30798913

fbshipit-source-id: 9b0702ca7515e538ad37310fd27633682c6bb820