improve results of type-at-pos on $ReadOnly types
Summary:
Improves type-at-pos etc. results around $ReadOnly annotations.

1/ The reason of the produced object used to overwrite the type alias reason
(if one existed) through the flag `invalidate_aliases:false`.
This reason is used by the normalizer to recover the name of the alias
(instead of falling back to the expanded version of the type). So to restore
this, we set this flag to `true`.

2/ The diff also uses the reason that corresponds to the entire
`$ReadOnly<T>` type for the result of the evaluated type, instead of the
reason for `T` that was used before.

3/ Finally, it adds a dedicated reason for the `$ReadOnly` portion which is
bypassed when creating the type of the returned object to improve some
of the resulting errors (ie. return the payload of the read-only, instead of
"$ReadOnly").

Changelog: [fix] improve results of type-at-pos on $ReadOnly types

Reviewed By: jbrown215

Differential Revision: D30467054

fbshipit-source-id: ccb57b421b3bd5a5d9d09b67b7b929732962ad43