restart on rebases in fs lazy mode
Summary:
when we detect a rebase, we compare how long we think it would take to recheck the working copy changes to how long we think it would take to restart and only have to recheck the changes since the new mergebase.

we can detect rebases when using watchman in hg repositories via the `hg.update` event. this _can_ work in either `fs` or `watchman` lazy modes, but we currently only do it for `watchman` mode.

in preparation for merging fs and watchman lazy modes into a single lazy mode, we can't branch differently for the two modes.

Changelog: [internal] see D30622710

Reviewed By: jbrown215

Differential Revision: D30615183

fbshipit-source-id: e4fded7ae860d18c21e8d331e8594c6ee40a7daa