merge lazy modes
Summary:
Replaces `fs` and `watchman` lazy modes with a single mode. Now that `fs` mode is SCM-aware and restarts on rebases, it is the same as `watchman` mode.

Changelog: [feat] The `ide`, `fs` and `watchman` lazy modes have been merged into a single mode. The new mode is now SCM-aware with or without Watchman; when the server starts, it queries Git or Mercurial and checks all existing local changes. Flowconfig setting migration:
- `lazy_mode=watchman` -> `lazy_mode=true`, `file_watcher=watchman`
- `lazy_mode=fs` -> `lazy_mode=true`
- `lazy_mode=ide` -> `lazy_mode=true`
- `lazy_mode=none` -> `lazy_mode=false`

Reviewed By: jbrown215

Differential Revision: D30622710

fbshipit-source-id: d9809c1a854a3afc399bf7ab2282d5031677c4cb