cleanup EUntypedImport and EUntypedTypeImport errors
Summary:
This diff pulls the code that handles EUntyped(Type)Import errors in flow_js.ml into a function
in Flow_js_utils.

It also adds a new constraint `CheckUntypedImportT` whose sole purpose is to handle
the case of
```
[declare] export [type] * from "source"
```
which does not use import/export related constraints otherwise.

Other than cleaning up Flow_js, this change enables the cleanup of D30922117, since
we do not rely on the call to `Import_export.mk_module_t` to raise EUntypedImport
errors for the above case. Earlier, this was probably a happy accident.

Reviewed By: mroch

Differential Revision: D30922116

fbshipit-source-id: 914db35c33f874caeb78b8adeda45fc79224dc15