Havoc variable state in old environment, and havoc variables with one assignment that is not at declaration
Summary: Variables shouldn't count as "constlike" if the one assignment to the variable is not at it's declaration. Also, havocing should change the initialization state of a variable. Without this, we get unsoundness like https://flow.org/try/#0G4QwTgBAHgXBB2BXAtgIwKZgNwCgBmi8AxgC4CWA9vBABYjAVEAUAlBAN44TfQQC8EACwAmXAF8cdBsxa4msCAzIATFkA. This diff implements this logic, but note that it doesn't change how havocing an uninitialized variable works in the new environment--havoced variables become fully initialized, not the equivalent of "MaybeInitialized" in the old env.

Reviewed By: panagosg7

Differential Revision: D29901504

fbshipit-source-id: 4002d3ad6a0645a2b2e9db2a95566126ea43169b