split flowLsp.state
Summary:
`state` is conflating two things: the client state (whether it's been initialized yet) and the server state (whether the LSP process is connected to a server).

all the server-related stuff only cares about whether it's connected or disconnected, so I split those two into `server_state`. not having to deal with `Pre_init` or `Post_shutdown` in a lot of places cleans up a lot of `option`'s and a few exceptions

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D28822815

fbshipit-source-id: 37d7613efc1ff69682078bbf723f03193d73d7c5