[new-env] Implement writing to new environment
Summary:
This diff implements variable writes in the new environment. There's two parts to writes: first, we find the entry in the var_env for the variable being written to, and unify its type with the incoming type.

Second, in order to constrain the types of variables by their annotations or providers, we look up the providers for the variable and flow the incoming type into the unioned types of the providers. This will result in the desired error if the variable is ill-typed.

Reviewed By: panagosg7

Differential Revision: D29858788

fbshipit-source-id: bd6aa13e87fd7b8e22b9e7f5fafad4488143816d