[new-env] Use new environment to power checking
Summary:
This is where we start to link everything together.

This diff makes `statement.ml` use the new environment module to handle checking when the new env is enabled. We do this by functorizing `statement`, `type_inference_js`, and many similar modules over `Env`, and providing the new env as an alternate implementation of the env signature. By doing this, we don't need to change the API used by statement.ml to access the environment much, other than passing in locations in some places where we didn't before. (Nearly all of the changes to `statement` in this diff are just the result of formatting changes by indenting the entire module one level.) When `type_inference_js`'s API is used, callers now need to choose between the old and new env implementations.

The meat of the actual changes in this diff is `new_env.ml` itself. The new env now has to do a number of things: it needs to use the env_builder to find writes for variables that the env builder is aware of, but it also needs to be able to handle internal variables, which don't exist from the env builder's perspective. It also needs to handle heap refinements (which at this point aren't handled by the env builder, though that will change), types (likewise), and contextual information about scope (e.g. are we currently at the module toplevel). Since the original env.ml already handles these things, the new env itself actually `include`s the entire implementation from the original env. It then shadows the definitions of functions relevant to the new env's behavior, e.g. `var_ref`, `set_var`, etc. In many cases, it needs to decide whether to use the new_env's infrastructure or the old env, by checking if the variable to be modified is internal or ordinary; that's why there's a lot of branching on the kind of Reason.name is passed in. (In cases where we don't branch on the name, it's because the name is a `string` and will always be an ordinary variable. That's an annoying inconsistency in the API but this isn't the place to solve it.)

Why not build a new implementation to handle internal vars etc? It sounded like a good idea, since it's pretty unwieldy to include the entire behavior of the old environment to handle a few internal variables and some state-tracking that could be handled elsewhere. However, it turns out that some internal variables really do depend on the complex behavior of the environment. For example, `super` and `this` are considered to be undeclared let-bound variables in derived class constructors, which are declared when `super()` is called, and accessing them (other than calling `super()`) before that is a TDZ error. `maybe_exhaustively_checked` also depends on the behavior of `merge_env`/`copy_env` to determine whether it's definitely initialized or not, which affects the result of exhaustiveness checks. Eventually, all of the internal variables will need to be handled of course, but they'll be handled in different ways--`maybe_exhaustively_checked` will probably go away, `this`/`super`/probably `exports` can be handled by the `env_builder`, and `return`/`yield` likely should be treated, not as variables, but as other unrelated contextual information (especially since they'll need to be locally inferred). So for now I think maintaining the original env's behavior is the best way to maintain support for these features while we work on the early stages of LTI.

All that said, with this diff we can now actually do some checking of programs using the new env, woohoo! This diff includes some test cases. There are differences between the new env's behavior here and the old env, as a result of combining provider-checks and open tvars. In particular, this part of `test.ml` causes an error:
```
var x = c ? 42 : "a'";

var y = 42

if (typeof x === 'string') {
  (x: string);
  y = x;
}
```
With the new env, when we assign `x` to `y`, we get an error. The assignment does a subtyping check against the type of the provider for `y`, which is 42, and since `x` is a string we blame the assignment. This is, of course, the desired behavior for us to eventually reach, but it's currently inconsistent--if `y`'s initializer was a function call instead, we wouldn't get the error, but `x` would instead flow into the function's tout:

```

var x2 = f(42); // where f : number => string
(x2: string); // <- error: number (from y below) is incompatible with string
x2 = y; // where y : number
```

With LTI, this will be fixed since the tout for the function will be resolved, and we'll get an error when we try to write to `x2`. However, to land this without LTI, the fix will be to have a separate tvar for the general type of a variable, which will happen in another diff.

Lastly, this diff creates a new_env mode for ./runtests, with -e. this results in the following:
`Passed: 148, Failed: 157, Skipped: 264, Errored: 159`
which is pretty decent for where we're at imo (it skips annotate_exports and friends)

Reviewed By: panagosg7

Differential Revision: D30616141

fbshipit-source-id: 890e7583ba4754d09771bdc3585a5dc4787e41e4