Filter out `any` types that have been propogated from React in codemods
Summary:
This diff filters out `any` types from unions where the type has been
propagated from React. React has some types that are marked as `any` that are
used when building components from JSX. This results in annotated components
often having props that are a union of `any | Props` where `Props` is the
actual prop type. Having a union with `any` negates the use of adding the type
annotation. Removing it from the union should result in better types overall.

Reviewed By: panagosg7

Differential Revision: D31003347

fbshipit-source-id: 66e8ad0aa0a910c3acacc530788db5c84c34c87a