Remove REF and improve havoc logic
Summary:
It turns out that havocing everything unconditionally when we enter function scopes isn't the right play. If a variable is constlike, then it's only written to once, which means that (modulo initialization) entering a function scope ought not to havoc anything. This is the logic for the old environment handled by D5386225 (https://github.com/facebook/flow/commit/e48030e66d653199cca8541e2f921ef95638b9ba); this diff applies the same logic in the env builder.

It also models the unsound-but-useful-and-sound-up-to-undefined logic that we assume that all variables are fully initialized inside functions. So, when a variable is uninitialized and we enter a function scope, we havoc it to its providers.

Finally, if a regular, non-function-entry havoc occurs and a variable is uninitialized, we shouldn't assume that it becomes definitely initialized after that. This diff implements that logic by keeping uninitialized on the val ref of a after a non-function-entry havoc if it was already uninitialized.

In order to accomplish the above goals, we need to be able to deeply inspect Val.t's. The REF write is an impediment to this, but it turns out its no longer needed, so this diff also removes REFs. This also lets us simplify `ssa` value structure a bit: instead of having a havoc consisting of a REF and a growing set of writes, we just have a single `havoc` Val.t. We do still need a handle on a definition `Loc.t` to test havocability, so that is directly encoded in the ssa value, passed in by bindings.

The deep inspection to see if a Val.t contains undefined is also a little complicated, because undefined can be refined away. Therefore, we use the `refinement_may_be_undefined` method to decide whether any `Refinement` write that deeply contains `Uninitialized` actually counts as containing uninitialized.

Reviewed By: jbrown215

Differential Revision: D29923271

fbshipit-source-id: 7bbaa416334c379a637c69c77dfbbd4526d557ff