fix suppressions in files that skip check and are not merged
Summary:
we have a recheck optimization that skips checking a file if none of its dependencies' signatures changed. but if we merged it, we added its suppressions but skip adding the errors during check, causing the suppressions to be unused. so we started deleting the suppressions for files we skip checking.

turns out, there are some dependents in `to_check` that we do not merge on this recheck. when we decide to skip rechecking them, we were incorrectly deleting their suppressions (regression in 0.160.0).

instead, we should only remove suppressions from files that we merge AND skip checking.

I suspect perhaps these files could just not be in `to_check` in the first place, but that is a more complex change to test and I want to unbreak 0.160.0.s

Reviewed By: panagosg7

Differential Revision: D31043400

fbshipit-source-id: 499404ae555aebe8f69b0c48090b2ed8f83c3725