Allow scope_builder to handle evals when not in flowmin-compatibility
Summary:
Currently the scope_builder fully ignores any function scope that contains a call to `eval` or a use of `with`. This is reasonable for flowmin purposes, but obviously is not compatible with the new environment. This diff fixes that, by allowing such scopes to be explored if `flowmin_compatibility` is disabled. This requires a little bit more threading of information through the ssa builder as well.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31040409

fbshipit-source-id: 67afbddc453f7a41b2b467e4cecece73121abd69