fix log spew about cancelled requests after server restarts
Summary:
when the server disconnects from the LSP process, the LSP process tells the client that all of the requests it was working on are cancelled. similarly, if the server isn't started yet, the LSP process will start one, but it will also cancel any incoming client requests. in either case, if the server is disconnected, all requests get cancelled.

if the client happens to cancel one of those requests as well, we currently log spew (and send a telemetry/event to the client). instead, we can just ignore the cancellation because it's a no-op.

Reviewed By: gkz

Differential Revision: D31073364

fbshipit-source-id: cfa4bd1c70b536fb0b953c117f6c34257d2392d9