add lazy mode suppression regression test
Summary:
adds a test of the regression in 0.160.1 that was caused by trying to drop suppressions for files that were merged but for which we skipped rechecking them.

this bug was reverted in D31103803 (https://github.com/facebook/flow/commit/0abc3bb2692271e77570bee1dccdac9844c26545) but we should still have a test to prevent it.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31101378

fbshipit-source-id: 195630da0dd0ea2f5ef7cd3988ce266e85d9d0be