support for SpecializeT in annotation inference
Summary:
This diff ports the logic associated with `SpecializeT` and `ThisSpecializeT`
uses to annotation inference:
* adds the relevant entries in the `elab_t` pattern match
* ports `instantiate_poly`, `fix_this_class` and helpers.

Worth noting here is that construction of annotations bypasses subtyping checks.
These checks do not have an effect in the end result of annotation inference
and are already performed during regular checking of the program.

Crucially, what enables this simplifying assumption is the lack of speculation
during annotation inference. Speculation would have required us to keep track
of subtyping errors for the purpose of backtracking.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D29422813

fbshipit-source-id: 411a518e08b421bbd1d3a5fcfa0f12cea58dc75b