mk_typeof_annotation in annotation inference
Summary:
Dealing with annotations is easy in annotation inference. Annot variables are
by definition 0->1, so there is no need to employ a mechanism like BecomeT
to enforce this.

Calling `mk_typeof_annotation` is essentially wrapping in an `AnnotT`
(for compatibility with the type structures produced in Flow_js) and adjusting
of reasons.

Reviewed By: mvitousek

Differential Revision: D29422807

fbshipit-source-id: 21690e43db102985fe16e2765c211d9286878d0d