Require variables to have providers
Summary:
As part of ensuring that variables are constrained by their providers, we need to make sure that all variables actually *do* have providers. This is dealt with separately here, because the error message to show when a variable doesn't have providers is different than when a variable is assigned to with a type incompatible with its providers.

This diff adds errors when variables are declared without an annotation or initializer, and then are never written to or are only written to within a nested function.

In the former case, this diff's change isn't strictly necessary for LTI--in a program like
```
var x;
x;
```
we know that `x` is always undefined, and both the new env and the old env handle it properly. However, this is still likely an error on the programmer's part and so it's useful to flag it. As an exception, though, we don't flag an error if the variable is never read from: a program like
```
var x;
```
alone won't raise an error. This is to minimize useless noise for the programmer when they're just starting to write a program, and haven't written the code that uses `x` yet.

If a variable *is* assigned to, but only in a nested scope, currently the provider analysis doesn't see that as a provider. In such a case, the programmer would need to annotate (but not necessarily initialize) the variable at its declaration site.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31042953

fbshipit-source-id: 43dd56d05bfebe7fd42d14c7dae05c8ea9a5d2e3