Constrain writes by providers
Summary:
This diff introduces an option to apply the new, provider-based constraints on variable writes to unannotated variables in the current environment. Every assignment or initialization now either populates the `loc_env` (which we reuse from the new environment as a map from locations to types) with the (not yet resolved) type of its RHS, if the write is a provider (determined by the find_providers analysis), or adds a post-inference subtyping check against the providers for the assigned-to variable if the write is not a provider.

This introduces a system configuration options for the  env to use, replacing both the `new_env` option and the totally unused old `check_updates_against_constraints` option. Currently, there's only one `env_option` that can be used to configure the classic env, but later diffs will add others.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D30970514

fbshipit-source-id: 9ca88bde7acdf848efd885e35bd3ff4faecc3ea0