Fix scopes and providers for class expressions
Summary:
It turns out that both the scope_builder (and by extension, the ssa and env builders), and the provider analysis, were wrong in the following case:
```
let x = class y { }
```
The scope builder never visited `y`, because the binding for `y` is a lexical binding and the lexical hoister ignores the RHSs of declarations and assignments. Meanwhile, the provider analysis would visit `y`, but would do so in the same scope as `x`, resulting in bindings for `x` and `y` coexisting in the same scope.

This diff fixes both systems, in both cases by creating a scope just for the class ID "in-between" the class body and the containing scope. This is analogous to, and implemented similarly to, how we handle `let f = function g() { ... }` (which *is* handled correctly). In the case of the scope_builder, we add the class' id to a Hoister bindings set before exploring this scope as well.

In order for these analyses to know whether we're looking at a class expression or a class declaration statement, we create separate methods in the ast mapper for each case (which normally just forward to the combined implementation, but which can be overridden separately to have specialized behavior like we need here).

Reviewed By: mroch

Differential Revision: D31100119

fbshipit-source-id: 4f2b213303ed3318f7760f7e105aca71e0819cb3