Require variables to have non-null providers
Summary:
Similar to the previous diff, we now also require that variables have providers that are not just null--if a variable is only ever null, then most likely the user failed to assign a non-null value to it.

In order to know whether a variable has providers that are non-null, we can see (internally to the provider_api) whether a provider's state is `NullInitialized`--if so, then we never saw a non-null provider for it. We include this information when deciding whether to error, and we only treat a variable as valid if it has providers and it's not just null initialized.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31062157

fbshipit-source-id: 87ab5bfeb92892b58b1b969929c96c2690370fa5