add $Partial utility type
Summary:
This adds a long-awaited utility type to Flow: $Partial. The semantics of this utility is to convert all of an object or interface's named fields to be optional, while maintaining all the object's other properties (exactness, sealedness, etc). This will also error if it is used with anything that is not interface-like.

See the added test for examples of the semantics of this new type.

[Changelog] : add $Partial utility type

Reviewed By: jbrown215

Differential Revision: D31173433

fbshipit-source-id: b8fc9bad7429ddcee91e48ba8b0436aa3d23f484