Add support for implicit this typing to LTI codemod
Summary:
This diff adds support to add annotations for implicit `this` typing errors. It works by matching the location of the error, then traversing into the function to look for a `this` using a finder. It then takes the inferred type of the `this` and creates an annotation to be added to the function with it.

This is done after annotating the function parameters to ensure that a function with unannotated parameters that also has an implicit `this` type will still be properly annotated.

Changelog: [internal]

Reviewed By: dsainati1

Differential Revision: D29856055

fbshipit-source-id: f385c29cfac08c8d9ee61b946759c7fc55cb7dae