make log saving timeouts specific to request types
Summary:
Changes the `log_saving_threshold_time_ms` flowconfig option to be a map from LSP request name to timeout threshold. It's a "multi" flowconfig option, like name reducers and excludes.

Configuring a few methods would look like this:
  log_saving=["textDocument/completion", 200]
  log_saving=["textDocument/hover", 500]

The syntax of the configuration is a JSON array with two elements where the first one is a string and the second one is a number. If we want to add more options, we can add an element to the array. I used JSON because this way we don't need to write a custom parser.

Reviewed By: mroch

Differential Revision: D31070186

fbshipit-source-id: ccc044fc7d0386bead41195b587620106f32c078