Ban reassigning functions and classes
Summary:
With LTI, we need to be able to treat annotated functions and class definitions as fixed typed values, in order to break reordering cycles push down context to contextually typed arguments etc. This diff implements that in the current environment, by simply applying the types-first rule against reassigned exported functions and classes to non-exported ones as well.

Changelog: [errors] Flow will now raise errors when function or class names are reassigned, even when not exported.

Reviewed By: panagosg7

Differential Revision: D30758221

fbshipit-source-id: f0817b0ece19894b2136eb64e5a55cd9533a29d5