Ban re-binding functions
Summary:
As in the previous diff, we want to be able to fully resolve variables eagerly when fully annotated, which is more difficult if they're re-declared. It's also very rarely a reasonable pattern to re-declare a function definition (since functions are hoisted anyways, so the last definition will always win--which I'm not sure if Flow actually models correctly now in the first place).

The exception is that non-predicate declared functions can be re-declared, either on their own or in combination with a regular function, in order to allow overloading. In order to make this distinction, this diff introduces a new binding kind for declared functions. Confusingly, prior to this diff, declared functions were `Var`s while regular functions were `Let (FunctionBinding)`s; now declared functions are `Let (DeclaredFunctionBinding)`s.

While I'm at it, this diff also bans re-declared parameters when the parameters are constlike.

Changelog: [internal] (previous diff change also covers this diff)

Reviewed By: panagosg7

Differential Revision: D30813376

fbshipit-source-id: 3705862ef2e45fc942c6f2f2738d983ec7b60536