remove babel from flow-dev-tools
Summary:
we were using babel 6 in flow-dev-tools, and in `newtests/` because flow-dev-tools' `tool test` is what runs those tests.

upgrading to babel 7 was nontrivial because it needs to apply the babel config to files outside of its own directory, and integrate with Facebook's internal build systems.

we don't really need babel transforms in modern node. we can remove babel and use flow-remove-types instead!

the only thing that's missing natively in node is ESM, which aren't enabled until node 14 and spew an "experimental" warning even then. going back to CJS isn't much of a loss.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31188238

fbshipit-source-id: 9c077cd86df46b0b22babe4c8547394050133f27