Allow variables to widen from null within nested function scopes
Summary:
Currently, the provider semantics only allow a provider to be in the same var scope as the variable was declared in. This was an issue in cases like the following:

```
declare var myList: Array<number>
var x; // or var x = null
myList.map((n) => { if (condition(n)) { x = n } }
```
Here, because `x` has no providers (or only a null provider) within its own car scope, we would consider it to be `void` or `null` and reject the assignment of `n`. However, this is a common enough patten in JS that we should support it if we can; and given that we now think we can initialize the environment without traversing the entire program, we should be able to handle this case.

The new rule for providers, implemented in this diff, is that we *prefer* providers from nearer scopes to the declaration of the variable, but we do allow providers from nested scopes. So while previously the rule was, when choosing which of two assignments to count as a provider, we select the one lexically earlier in the program as long as it's in the same scope as the variable, now its:
1. choose the one in the nearer var scope to the variable's declaration, and
2. if they are in the same var scope, choose the lexically earlier one..

This means that in a program like the following:
```
var x;
function f() { x = 42; }
x = "a"
```
we will consider the provider for `x` to be "a", and when we constrain against writes we'll reject the assignment `x = 42`.

We're still allowed to progress from uninitialized to null-initialized to initialized, and that happens in the same nearer-scope-first fashion.
```
var y;
function f() { y = 42; }
y = null;
```
Here, both 42 and null are providers for `y`. Compare to:
```
var z;
function f() { z = null; }
z = 42;
```
Here, the only provider for `z` is 42 and the null assignment is rejected--again, because we went from uninitialized to initialized in `z`'s declaration scope, which "beats" the assignment of null in the nested scope.
 ---
This diff implements this work in the provider analysis. It requires a lot of additional bookkeeping, mainly because the order in which the AST traversal explores the program is different than the semantic order in which we consider initialization: the AST traversal explores top-down, depth-first, while the semantics put forth above are basically breadth-first. We handle this by including, in the `state` type that tracks the initialization state of a variable and the `write_state` type that tracks what is happening with a single, individual assignment to that variable, the scope depth at which the variable was initialized. We then use that to ensure that closer-in writes "beat" deeper writes, even if the deeper write was explored first in the analysis.

So, tracking what happens in two programs:
```
var w;
// w starts off as `Uninitialized`

function f() { w = 42; }
// w gets an incoming `Value` write with depth 1.
// `w`'s state moves to `Initialized (depth=1, null_depth=None)`;
// the `null_depth = None part is because `w` has not been initialized by `null`.

w = null;
// w gets an incoming `Null` write with depth 0.
// `w`'s state moves to `Initialized (depth=1, null_depth=0), indicating that
// the null initializer "counts" as a provider, since it was at a shallowed depth
// than the full initialization.
```
compare to
```
var z;
// z starts off as `Uninitialized`

function f() { z = null; }
// z gets an incoming `Null` write with depth 1. `z`'s state moves to `NullInitialized (depth=1)`.

z = 42;
// z gets an incoming `Value` write with depth 0. Since 0 < 1, we no
// longer consider `z` to have `null` as a provider
// and it moves to `Initialized (depth=0, null_depth=None)
```

As we go, we add *potential* providers (i.e., writes which result in the variable state changing) to the variable's `provider_locs,` which is now here a map from locs to write_states. At the end, we can compare the provider entries to the overall variable state and filter out any potential providers that we later determined weren't providers based on the final overall variable state--for example, `z = null` above is recorded as a provider with the write_state (`Null (depth=1)`), but since the final
state of the variable is `Initialized (depth=0, null_depth=None)`, and the `null_depth=None` part indicates that there were no null initializers in the right place to count as a provider, we'll remove that write from the providers for `z` in `Provider_api.simplify`.

Reviewed By: panagosg7

Differential Revision: D31077258

fbshipit-source-id: 2e83e27f86cf3ae7b80b2478a21322afa1f66ce5